Add python_requires metadata.
Once the package is published this inform pip about which version of
Python this package is compatible with. Without that it could be
installed and failed at import time, which is less than ideal.

In particular if you are Python  only, think about adding your project
to https://python3statement.org/,